Refactored awsExplorer initialization into an activation file
