+++ .gitignore
@@ -9,7 +9,7 @@ node_modules
 telemetryCache
 quickStart.html
 .gitcommit
-media/libs/lodash.min.js
+media/libs/
 
 # Auto generated Service Clients
 #TODO: remove-when-schema-sdk-launched
