+++ src/test/awsExplorer/awsExplorer.test.ts
@@ -6,31 +6,19 @@
 import * as assert from 'assert'
 import { AwsExplorer } from '../../awsexplorer/awsExplorer'
 import { RegionNode } from '../../awsexplorer/regionNode'
-import { AwsContextTreeCollection } from '../../shared/awsContextTreeCollection'
-import { MockOutputChannel } from '../mockOutputChannel'
 import {
     DEFAULT_TEST_REGION_CODE,
     DEFAULT_TEST_REGION_NAME,
     FakeAwsContext,
-    FakeRegionProvider,
-    FakeResourceFetcher
+    FakeRegionProvider
 } from '../utilities/fakeAwsContext'
 
 describe('AwsExplorer', () => {
     it('displays region nodes with user-friendly region names', async () => {
         const awsContext = new FakeAwsContext()
         const regionProvider = new FakeRegionProvider()
-        const awsContextTreeCollection = new AwsContextTreeCollection()
-        const resourceFetcher = new FakeResourceFetcher()
-        const mockChannel = new MockOutputChannel()
 
-        const awsExplorer = new AwsExplorer(
-            awsContext,
-            awsContextTreeCollection,
-            regionProvider,
-            resourceFetcher,
-            mockChannel
-        )
+        const awsExplorer = new AwsExplorer(awsContext, regionProvider)
 
         const treeNodesPromise = awsExplorer.getChildren()
